+++ web/modules/custom/hoeringsportal_citizen_proposal/src/Helper/Helper.php
@@ -1,0 +1,100 @@
+<?php
+
+namespace Drupal\hoeringsportal_citizen_proposal\Helper;
+
+use Drupal\Core\State\State;
+use Drupal\Core\StringTranslation\StringTranslationTrait;
+use Drupal\Core\TempStore\PrivateTempStore;
+use Drupal\Core\TempStore\PrivateTempStoreFactory;
+use Drupal\Core\Url;
+use Drupal\node\Entity\Node;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\Serializer\Serializer;
+use Drupal\Core\Messenger\MessengerTrait;
+
+/**
+ * A helper class for the module.
+ */
+class Helper {
+  use MessengerTrait;
+  use StringTranslationTrait;
+
+  private const CITIZEN_PROPOSAL_ENTITY = 'citizen_proposal_entity';
+
+  /**
+   * Constructor for the citizen proposal helper class.
+   */
+  public function __construct(
+    readonly private PrivateTempStoreFactory $tempStoreFactory,
+    readonly private Serializer $serializer,
+    readonly private State $state,
+  ) {
+  }
+
+  /**
+   * Validate that the tempstore contains a node and return it.
+   *
+   * @return object|null
+   *   The entity or NULL if no valid entity was found.
+   */
+  public function getDraftProposal(): ?object {
+    $entitySerialized = $this->getProposalStorage()->get(self::CITIZEN_PROPOSAL_ENTITY);
+    $node = is_string($entitySerialized) ? $this->serializer->deserialize($entitySerialized, Node::class, 'json') : NULL;
+
+    return isset($node) && $node instanceof Node ? $node : NULL;
+  }
+
+  /**
+   * Delete the citizen_proposal_entity tempstore.
+   */
+  public function deleteDraftProposal(): void {
+    try {
+      $this->getProposalStorage()->delete(self::CITIZEN_PROPOSAL_ENTITY);
+    }
+    catch (\Exception $e) {
+    }
+  }
+
+  /**
+   * Add entity to temp store.
+   */
+  public function setDraftProposal($entity): void {
+    $nodeSerialized = $this->serializer->serialize($entity, 'json');
+    try {
+      $this->getProposalStorage()->set(self::CITIZEN_PROPOSAL_ENTITY, $nodeSerialized);
+    }
+    catch (\Exception $e) {
+    }
+  }
+
+  /**
+   * Abandon submission and add redirect response.
+   *
+   * @return \Symfony\Component\HttpFoundation\RedirectResponse
+   *   A redirect response.
+   */
+  public function abandonSubmission(): RedirectResponse {
+    $this->messenger()->addWarning($this->t('Could not find a proposal to approve.'));
+    $url = Url::fromRoute('hoeringsportal_citizen_proposal.citizen_proposal.proposal_add');
+
+    return new RedirectResponse($url->toString());
+  }
+
+  /**
+   * Preprocess forms related to citizen proposal.
+   */
+  public function preprocessForm(&$variables): void {
+    $variables['admin_form_state_values'] = $this->state->get('citizen_proposal_admin_form_values');
+  }
+
+  /**
+   * The proposal temp store.
+   *
+   * @return \Drupal\Core\TempStore\PrivateTempStore
+   *   The proposal tempstore.
+   */
+  private function getProposalStorage(): PrivateTempStore {
+    return $this->tempStoreFactory->get('hoeringsportal_citizen_proposal');
+  }
+
+}